+++ odoo/local-src/stock_putaway_route/__manifest__.py
@@ -30,6 +30,7 @@
         ],
     'data': [
         'views/stock.xml',
+        'security/ir.model.access.csv',
     ],
     'installable': True,
     'auto_install': False,