+++ odoo/local-src/specific_sale/__manifest__.py
@@ -13,6 +13,7 @@
         'sale',
         'sale_exception',
         'sale_product_additional',
+        'stock',
         'stock_available_immediately',
     ],
     'website': 'http://www.camptocamp.com',