+++ odoo/local-src/specific_sale/models/__init__.py
@@ -5,3 +5,4 @@
 from . import product_template
 from . import res_partner
 from . import sale
+from . import stock_production_lot