+++ HISTORY.rst
@@ -12,6 +12,8 @@ Unreleased
 * ALCN-788: Set the reference with the supplier invoice number
 * ALCN-794: Set the flag "update_posted" on several journals and set the field check total required on vendor bills.
 
+* Logistics: Reception: fix set default lot name with date, added some help, reception wizard (added button to allow to move to next destination)
+
 **Bugfixes**
 * ALCN-812: Add missing ir.model.access